requirements.txt
langchain-google-genai
langchain-openai
langchain-anthropic
langchain-ollama
python-dotenv
gradio>=4.0.0
langchain
langchain-community